Fix seed calculation for random number generator
P::tstep is uint, we want to maybe convert it to uint64 before doing the calculation(?)

+++ vlasovsolver/vlasovmover.cpp
@@ -424,7 +424,7 @@ void calculateAcceleration(const uint popID,const uint globalMaxSubcycles,const 
    // set seed, initialise generator and get value. The order is the same
    // for all cells, but varies with timestep.
    std::knuth_b rndState;
-   uint64_t seed=(P::seed*(P::tstep+1)<<32);
+   uint64_t seed=(P::seed*(uint64_t(P::tstep)+1)<<32);
    rndState.seed(seed);    
    uint map_order = std::uniform_int_distribution<>(0,2)(rndState);
